Added ArchiveFolder configuration

+++ src/DWC_A/ArchiveFolder.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using DwC_A.Config;
+using System;
 using System.IO;
 using System.IO.Compression;
 
@@ -8,6 +9,7 @@ namespace DwC_A
     {
         private readonly string fileName;
         private readonly string folderPath;
+        private readonly ArchiveFolderConfiguration config;
         
         public bool ShouldCleanup { get; private set; }
 
@@ -15,11 +17,12 @@ namespace DwC_A
         /// Extracts archive to a folder
         /// </summary>
         /// <param name="fileName">Zip archive file name</param>
-        /// <param name="folderPath">Path to extract to.  Leave null for a temp folder</param>
-        public ArchiveFolder(string fileName, string folderPath = null)
+        /// <param name="config"><see cref="ArchiveFolderConfiguration"/></param>
+        public ArchiveFolder(string fileName, ArchiveFolderConfiguration config)
         {
+            this.config = config;
             this.fileName = fileName;
-            this.folderPath = string.IsNullOrEmpty(folderPath) ? GetTempPath() : folderPath;
+            this.folderPath = string.IsNullOrEmpty(config.OutputPath) ? GetTempPath() : config.OutputPath;
         }
 
         public string Extract()
@@ -30,7 +33,10 @@ namespace DwC_A
 
         public void DeleteFolder()
         {
-            Directory.Delete(folderPath, true);
+            if (config.ShouldCleanup)
+            {
+                Directory.Delete(folderPath, true);
+            }
         }
 
         private string GetTempPath()
